refactor: remove schedule templates feature
Schedule templates duplicated the copy-day workflow. Drop the
settings page, components, server actions, DB table, and i18n
keys; relocate copy-day server actions to their own file.

Closes #110
