--- app/[tenant]/admin/settings/templates/page.tsx
@@ -1,15 +1,0 @@
-import { getTranslations } from 'next-intl/server'
-import { requireTenantEditor } from '@/lib/guards'
-import { ScheduleTemplateManagement } from '@/components/schedule-template-management'
-
-export default async function TemplatesSettingsPage() {
-  await requireTenantEditor()
-  const t = await getTranslations('Tenant.settings')
-
-  return (
-    <div className="mx-auto max-w-3xl px-6 py-8">
-      <h1 className="mb-6 text-2xl font-semibold">{t('tabTemplates')}</h1>
-      <ScheduleTemplateManagement />
-    </div>
-  )
-}
