+++ components/copy-day-dialog.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from 'react'
 import { useRouter } from 'next/navigation'
 import { useTranslations } from 'next-intl'
 import { toast } from 'sonner'
-import { copyDaySections } from '@/app/actions/schedule-templates'
+import { copyDaySections } from '@/app/actions/copy-day'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
