+++ messages/de.json
@@ -128,8 +128,7 @@
       "addBreakfast": "Frühstück hinzufügen",
       "addActivity": "Aktivität",
       "addReservation": "Reservierung hinzufügen",
-      "copyDay": "Kopieren",
-      "templates": "Vorlagen"
+      "copyDay": "Kopieren"
     },
     "handover": {
       "toggle": "Übergabe",
@@ -372,7 +371,6 @@
       "tabLanguage": "Sprache",
       "tabMembers": "Team",
       "tabStaff": "Personal",
-      "tabTemplates": "Vorlagen",
       "tabFeedback": "Feedback",
       "languageTitle": "Sprache des Betriebs",
       "languageDescription": "Diese Sprache wird für alle Mitglieder Ihres Betriebs verwendet.",
@@ -526,57 +524,6 @@
       "copiedActivities": "Aktivitäten kopiert.",
       "copiedShifts": "Dienstplan kopiert.",
       "copiedBoth": "Aktivitäten und Dienstplan kopiert."
-    },
-    "templates": {
-      "dialogTitle": "Zeitplanvorlagen",
-      "dialogDescription": "Gespeichertes Programm auf diesen Tag anwenden oder die aktuellen Aktivitäten als wiederverwendbare Vorlage speichern (unter Einstellungen → Vorlagen).",
-      "pageIntro": "Zeitplanvorlagen sind benannte, wiederverwendbare Aktivitätslisten (Zeiten, Gästezahl, Veranstaltungstyp, Ansprechpartner). So setzen Sie ein typisches Tagesprogramm schnell auf beliebige Daten.",
-      "pageVsCopyDay": "Das ist etwas anderes als „Auf einen anderen Tag kopieren“: Dort wird ein konkreter Tag 1:1 an ein Datum dupliziert. Vorlagen nutzen Sie für wiederkehrende Muster; Kopieren, wenn ein echter Ablauf übernommen werden soll.",
-      "listHeading": "{count, plural, =0 {Keine Vorlagen} one {# Vorlage} other {# Vorlagen}}",
-      "newTemplate": "Neue Vorlage",
-      "edit": "Bearbeiten",
-      "editorCreateTitle": "Neue Zeitplanvorlage",
-      "editorEditTitle": "Zeitplanvorlage bearbeiten",
-      "editorSave": "Speichern",
-      "editorSaving": "Wird gespeichert\u2026",
-      "templateCreated": "Vorlage erstellt.",
-      "templateUpdated": "Vorlage aktualisiert.",
-      "addActivityLine": "Aktivität hinzufügen",
-      "moveUpAria": "Nach oben",
-      "moveDownAria": "Nach unten",
-      "removeLineAria": "Aktivität entfernen",
-      "activitiesSection": "Aktivitäten in dieser Vorlage",
-      "loadingMeta": "Lade Veranstaltungstypen und Kontakte\u2026",
-      "noneOption": "Keine",
-      "needOneActivity": "Mindestens eine Aktivität mit Titel hinzufügen.",
-      "nameRequired": "Vorlagenname erforderlich.",
-      "tabApply": "Vorlage anwenden",
-      "tabSave": "Als Vorlage speichern",
-      "loading": "Laden\u2026",
-      "noTemplates": "Noch keine gespeicherten Vorlagen.",
-      "templateLabel": "Vorlage",
-      "selectTemplate": "Vorlage auswählen\u2026",
-      "activityCount": "{count, plural, one {# Aktivität} other {# Aktivitäten}}",
-      "modeLabel": "Modus",
-      "modeMerge": "Zusammenführen (zu vorhandenen hinzufügen)",
-      "modeReplace": "Alle ersetzen",
-      "cancel": "Abbrechen",
-      "apply": "Anwenden",
-      "applying": "Wird angewendet\u2026",
-      "applied": "Vorlage angewendet.",
-      "noActivities": "Keine Aktivitäten zum Speichern als Vorlage.",
-      "saveCount": "{count, plural, one {# Aktivität} other {# Aktivitäten}} als Vorlage speichern.",
-      "nameLabel": "Vorlagenname",
-      "namePlaceholder": "z. B. Wochenendzeitplan",
-      "saveButton": "Vorlage speichern",
-      "saving": "Wird gespeichert\u2026",
-      "saved": "Vorlage gespeichert.",
-      "settingsEmpty": "Noch keine Vorlagen. Erstellen Sie eine hier, oder speichern Sie in der Tagesansicht unter Vorlagen → Als Vorlage speichern.",
-      "deleteTitle": "Vorlage löschen?",
-      "deleteDescription": "\u201e{name}\u201c wird dauerhaft gelöscht.",
-      "delete": "Löschen",
-      "deleting": "Wird gelöscht\u2026",
-      "deleted": "Vorlage gelöscht."
     },
     "staff": {
       "section": {
