+++ messages/en.json
@@ -232,8 +232,7 @@
       "addBreakfast": "Add breakfast",
       "addActivity": "Activity",
       "addReservation": "Add reservation",
-      "copyDay": "Copy",
-      "templates": "Templates"
+      "copyDay": "Copy"
     },
     "handover": {
       "toggle": "Hand-over",
@@ -500,7 +499,6 @@
       "tabLanguage": "Language",
       "tabMembers": "Team",
       "tabStaff": "Staff",
-      "tabTemplates": "Templates",
       "tabFeedback": "Feedback",
       "languageTitle": "Venue language",
       "languageDescription": "This language will be used for all members of your venue.",
@@ -654,58 +652,6 @@
       "copiedActivities": "Activities copied.",
       "copiedShifts": "Staff schedule copied.",
       "copiedBoth": "Activities and staff schedule copied."
-    },
-    "templates": {
-      "dialogTitle": "Schedule templates",
-      "dialogDescription": "Load a saved programme onto this day, or save the current day’s activities as a template you can reuse from Settings → Templates.",
-      "pageIntro": "Schedule templates are named, reusable activity lists (times, covers, venue, point of contact). Use them to drop a standard day’s programme onto any date in a couple of clicks.",
-      "pageVsCopyDay": "This is not the same as “copy to another day” on the day view, which duplicates one specific day onto another date. Use templates when you want a repeatable pattern; use copy when you are cloning a real day as it was run.",
-      "listHeading": "{count, plural, =0 {No templates} one {# template} other {# templates}}",
-      "newTemplate": "New template",
-      "edit": "Edit",
-      "editorCreateTitle": "New schedule template",
-      "editorEditTitle": "Edit schedule template",
-      "editorSave": "Save",
-      "editorSaving": "Saving…",
-      "templateCreated": "Template created.",
-      "templateUpdated": "Template updated.",
-      "addActivityLine": "Add activity",
-      "moveUpAria": "Move up",
-      "moveDownAria": "Move down",
-      "removeLineAria": "Remove activity",
-      "activitiesSection": "Activities in this template",
-      "loadingMeta": "Loading venue types and contacts…",
-      "noneOption": "None",
-      "needOneActivity": "Add at least one activity with a title.",
-      "nameRequired": "Template name is required.",
-      "tabApply": "Apply template",
-      "tabSave": "Save as template",
-      "loading": "Loading…",
-      "noTemplates": "No saved templates yet.",
-      "templateLabel": "Template",
-      "selectTemplate": "Select a template…",
-      "activityCount": "{count, plural, one {# activity} other {# activities}}",
-      "modeLabel": "Mode",
-      "modeMerge": "Merge (add to existing)",
-      "modeReplace": "Replace all",
-      "cancel": "Cancel",
-      "apply": "Apply",
-      "applying": "Applying…",
-      "applied": "Template applied.",
-      "noActivities": "No activities to save as template.",
-      "saveCount": "Saving {count, plural, one {# activity} other {# activities}} as a template.",
-      "nameLabel": "Template name",
-      "namePlaceholder": "e.g. Weekend schedule",
-      "saveButton": "Save template",
-      "saving": "Saving…",
-      "saved": "Template saved.",
-      "settingsEmpty": "You have no templates yet. Create one here, or on any day use Templates → Save as template to capture that day’s programme.",
-      "deleteTitle": "Delete template?",
-      "deleteDescription": "“{name}” will be permanently deleted.",
-      "delete": "Delete",
-      "deleting": "Deleting…",
-      "deleted": "Template deleted.",
-      "goToDayView": "Open today's day view"
     },
     "staff": {
       "section": {
