+++ messages/es.json
@@ -128,8 +128,7 @@
       "addBreakfast": "Añadir desayuno",
       "addActivity": "Actividad",
       "addReservation": "Añadir reserva",
-      "copyDay": "Copiar",
-      "templates": "Plantillas"
+      "copyDay": "Copiar"
     },
     "handover": {
       "toggle": "Entrega de turno",
@@ -372,7 +371,6 @@
       "tabLanguage": "Idioma",
       "tabMembers": "Equipo",
       "tabStaff": "Personal",
-      "tabTemplates": "Plantillas",
       "tabFeedback": "Sugerencias",
       "languageTitle": "Idioma del establecimiento",
       "languageDescription": "Este idioma se usará para todos los miembros de tu establecimiento.",
@@ -526,57 +524,6 @@
       "copiedActivities": "Actividades copiadas.",
       "copiedShifts": "Horario de personal copiado.",
       "copiedBoth": "Actividades y horario de personal copiados."
-    },
-    "templates": {
-      "dialogTitle": "Plantillas de horario",
-      "dialogDescription": "Cargar un programa guardado en este día, o guardar las actividades actuales como plantilla reutilizable (Configuración → Plantillas).",
-      "pageIntro": "Las plantillas de horario son listas de actividades con nombre y reutilizables (horas, cubiertos, tipo de espacio, contacto). Úsalas para colocar un programa tipo en cualquier fecha en pocos clics.",
-      "pageVsCopyDay": "No es lo mismo que « Copiar a otro día »: eso duplica un día concreto en una fecha. Usa plantillas para un patrón repetible; la copia cuando quieras clonar un día tal como se hizo.",
-      "listHeading": "{count, plural, =0 {Sin plantillas} one {# plantilla} other {# plantillas}}",
-      "newTemplate": "Nueva plantilla",
-      "edit": "Editar",
-      "editorCreateTitle": "Nueva plantilla de horario",
-      "editorEditTitle": "Editar plantilla de horario",
-      "editorSave": "Guardar",
-      "editorSaving": "Guardando\u2026",
-      "templateCreated": "Plantilla creada.",
-      "templateUpdated": "Plantilla actualizada.",
-      "addActivityLine": "Añadir actividad",
-      "moveUpAria": "Subir",
-      "moveDownAria": "Bajar",
-      "removeLineAria": "Quitar actividad",
-      "activitiesSection": "Actividades en esta plantilla",
-      "loadingMeta": "Cargando tipos de espacio y contactos\u2026",
-      "noneOption": "Ninguno",
-      "needOneActivity": "Añade al menos una actividad con título.",
-      "nameRequired": "El nombre de la plantilla es obligatorio.",
-      "tabApply": "Aplicar plantilla",
-      "tabSave": "Guardar como plantilla",
-      "loading": "Cargando\u2026",
-      "noTemplates": "Ninguna plantilla guardada aún.",
-      "templateLabel": "Plantilla",
-      "selectTemplate": "Seleccionar una plantilla\u2026",
-      "activityCount": "{count, plural, one {# actividad} other {# actividades}}",
-      "modeLabel": "Modo",
-      "modeMerge": "Combinar (añadir a las existentes)",
-      "modeReplace": "Reemplazar todo",
-      "cancel": "Cancelar",
-      "apply": "Aplicar",
-      "applying": "Aplicando\u2026",
-      "applied": "Plantilla aplicada.",
-      "noActivities": "Ninguna actividad para guardar como plantilla.",
-      "saveCount": "Guardando {count, plural, one {# actividad} other {# actividades}} como plantilla.",
-      "nameLabel": "Nombre de la plantilla",
-      "namePlaceholder": "ej. Horario de fin de semana",
-      "saveButton": "Guardar plantilla",
-      "saving": "Guardando\u2026",
-      "saved": "Plantilla guardada.",
-      "settingsEmpty": "Aún no hay plantillas. Crea una aquí, o en un día: Plantillas → Guardar como plantilla.",
-      "deleteTitle": "¿Eliminar plantilla?",
-      "deleteDescription": "\u201c{name}\u201d se eliminará permanentemente.",
-      "delete": "Eliminar",
-      "deleting": "Eliminando\u2026",
-      "deleted": "Plantilla eliminada."
     },
     "staff": {
       "section": {
