+++ messages/fr.json
@@ -232,8 +232,7 @@
       "addBreakfast": "Ajouter petit-déjeuner",
       "addActivity": "Activité",
       "addReservation": "Ajouter une réservation",
-      "copyDay": "Copier",
-      "templates": "Modèles"
+      "copyDay": "Copier"
     },
     "handover": {
       "toggle": "Passation",
@@ -500,7 +499,6 @@
       "tabLanguage": "Langue",
       "tabMembers": "Équipe",
       "tabStaff": "Personnel",
-      "tabTemplates": "Modèles",
       "tabFeedback": "Retours",
       "languageTitle": "Langue de l'établissement",
       "languageDescription": "Cette langue sera utilisée pour tous les membres de votre établissement.",
@@ -654,58 +652,6 @@
       "copiedActivities": "Activités copiées.",
       "copiedShifts": "Planning du personnel copié.",
       "copiedBoth": "Activités et planning du personnel copiés."
-    },
-    "templates": {
-      "dialogTitle": "Modèles de planning",
-      "dialogDescription": "Charger un programme enregistré sur ce jour, ou enregistrer les activités actuelles comme modèle réutilisable (Paramètres → Modèles).",
-      "pageIntro": "Les modèles de planning sont des listes d'activités nommées et réutilisables (horaires, couverts, type de lieu, contact). Servez-vous en pour placer un programme type sur n'importe quelle date en quelques clics.",
-      "pageVsCopyDay": "Ce n'est pas la même chose que « Copier vers un autre jour » : cette action duplique un jour précis vers une date. Utilisez les modèles pour un schéma répétable ; la copie quand il faut reproduire une journée telle qu'elle a eu lieu.",
-      "listHeading": "{count, plural, =0 {Aucun modèle} one {# modèle} other {# modèles}}",
-      "newTemplate": "Nouveau modèle",
-      "edit": "Modifier",
-      "editorCreateTitle": "Nouveau modèle de planning",
-      "editorEditTitle": "Modifier le modèle de planning",
-      "editorSave": "Enregistrer",
-      "editorSaving": "Enregistrement…",
-      "templateCreated": "Modèle créé.",
-      "templateUpdated": "Modèle mis à jour.",
-      "addActivityLine": "Ajouter une activité",
-      "moveUpAria": "Monter",
-      "moveDownAria": "Descendre",
-      "removeLineAria": "Retirer l'activité",
-      "activitiesSection": "Activités dans ce modèle",
-      "loadingMeta": "Chargement des types de lieu et des contacts…",
-      "noneOption": "Aucun",
-      "needOneActivity": "Ajoutez au moins une activité avec un libellé.",
-      "nameRequired": "Le nom du modèle est requis.",
-      "tabApply": "Appliquer un modèle",
-      "tabSave": "Enregistrer comme modèle",
-      "loading": "Chargement…",
-      "noTemplates": "Aucun modèle enregistré pour l'instant.",
-      "templateLabel": "Modèle",
-      "selectTemplate": "Sélectionner un modèle…",
-      "activityCount": "{count, plural, one {# activité} other {# activités}}",
-      "modeLabel": "Mode",
-      "modeMerge": "Fusionner (ajouter aux existantes)",
-      "modeReplace": "Remplacer tout",
-      "cancel": "Annuler",
-      "apply": "Appliquer",
-      "applying": "Application…",
-      "applied": "Modèle appliqué.",
-      "noActivities": "Aucune activité à enregistrer comme modèle.",
-      "saveCount": "Enregistrement de {count, plural, one {# activité} other {# activités}} comme modèle.",
-      "nameLabel": "Nom du modèle",
-      "namePlaceholder": "ex. Planning du week-end",
-      "saveButton": "Enregistrer le modèle",
-      "saving": "Enregistrement…",
-      "saved": "Modèle enregistré.",
-      "settingsEmpty": "Aucun modèle pour l’instant. Créez-en un ici, ou sur un jour : Modèles → Enregistrer comme modèle.",
-      "deleteTitle": "Supprimer le modèle ?",
-      "deleteDescription": "« {name} » sera définitivement supprimé.",
-      "delete": "Supprimer",
-      "deleting": "Suppression…",
-      "deleted": "Modèle supprimé.",
-      "goToDayView": "Ouvrir la vue du jour"
     },
     "staff": {
       "section": {
